[BEAM-1925] validate DoFn at pipeline creation time
