+++ sdks/python/setup.py
@@ -132,6 +132,7 @@ setuptools.setup(
         'apache_beam/runners/worker/opcounters.py',
         'apache_beam/runners/worker/operations.py',
         'apache_beam/transforms/cy_combiners.py',
+        'apache_beam/transforms/dofn_signature.py',
         'apache_beam/utils/counters.py',
         'apache_beam/utils/windowed_value.py',
     ]),
